+++ Library/simulink/callback/ThreePhaseConstantPowerSourceLoad_cb.m
@@ -1,12 +1,12 @@
-%% CONSTANTPOWERSOURCELOAD_CB - Implements callbacks for 'Constant Power
-% Source or Load' block
+%% ThreePhaseConstantPowerSourceLoad_cb - Implements callbacks for 
+% 'Three-Phase Constant Power Source or Load' block
 %
-% This function implements the mask callbacks for the 'Constant Power
-% Source or Load' block in the NREL Campus Energy Modeling Simulink block
-% library. It is designed to be called from the block mask.
+% This function implements the mask callbacks for the 'Three-Phase Constant
+% Power Source or Load' block in the NREL Campus Energy Modeling Simulink
+% block library. It is designed to be called from the block mask.
 % 
 % SYNTAX:
-%   varargout = ConstantPowerSourceLoad_cb(block, callback, varargin)
+%   varargout = ThreePhaseConstantPowerSourceLoad_cb(block, callback, varargin)
 %
 % INPUTS:
 %   block =     Simulink block path
@@ -21,7 +21,8 @@
 %    Energy Modeling Simulink library; therefore the error checking and
 %    documentation are minimal. View the code to see what is going on.
 
-function varargout = ConstantPowerSourceLoad_cb(block, callback, varargin)
+function varargout = ThreePhaseConstantPowerSourceLoad_cb(block, ...
+    callback, varargin)
 
     %% Setup
     % Default output = none
@@ -33,7 +34,8 @@ function varargout = ConstantPowerSourceLoad_cb(block, callback, varargin)
         % Initialization
         case 'init'
             varargout = ...
-                ConstantPowerSourceLoad_cb_init(block, varargin{:});
+                ThreePhaseConstantPowerSourceLoad_cb_init(block, ...
+                    varargin{:});
             
         otherwise
             warning([block ':unimplementedCallback'], ...
@@ -44,19 +46,22 @@ end
 
 %% Subfunctions
 % Initialization
-function out = ConstantPowerSourceLoad_cb_init(block, Conven, ...
+function out = ThreePhaseConstantPowerSourceLoad_cb_init(block, Conven, ...
     ~, ~, VNom, fNom, VLow)
     % Check voltages
     assert(VNom > 0, [block ':incorrectParameter'], ...
-        ['Nominal voltage for a constant power source/load block ' ...
+        ['Nominal line-to-line voltage for a ' ...
+         'three-phase constant power source/load block ' ...
          'must be strictly positive.']);
     assert(VLow > 0, [block ':incorrectParameter'], ...
-        ['Lower cutoff voltage for a constant power source/load block ' ...
+        ['Lower cutoff voltage for a ' ...
+         'three-phase constant power source/load block ' ...
          'must be strictly positive.']);
      
     % Check frequency
     assert(fNom > 0, [block ':incorrectParameter'], ...
-        ['Nominal frequency for a constant power source/load block ' ...
+        ['Nominal frequency for a ' ...
+         'three-phase constant power source/load block ' ...
          'must be strictly positive.']);
     
     % Note: block internals are with a load convention.